+++ .eslintrc.js
@@ -11,5 +11,14 @@ module.exports = {
     ecmaVersion: 'latest'
   },
   rules: {
+    // Customize eslint standard style definitions
+    'comma-dangle': ['error', 'never'],
+    'linebreak-style': ['error', 'unix'],
+    'max-len': ['error', { code: 100, tabWidth: 2, ignoreUrls: true }],
+    semi: ['error', 'always'],
+    'object-shorthand': 'off'
+
+    // some variables are unused to keep template symmetrical
+    // 'no-unused-vars': 'off'
   }
-}
+};