+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to
 ## v0.0.8-dev 2023-07-02
 
 - Bump dotenv@16.3.1, node-fetch@2.6.12 to clean npm outdated warnings.
-- Re-install eslint, manual install semver@7.5.3, delete and regenerate package-lock.json to clear npm audit warning.
+- Re-install eslint, manual install semver@7.5.3, delete and regenerate package-lock.json in v3 format to clear npm audit warning.
 
 ## [v0.0.7](https://github.com/cotarr/collab-iot-device/releases/tag/v0.0.7) - 2023-01-11